[unrated] Title: 크기가 작은 부분문자열, Time: 1.71 ms, Memory: 76.2 MB -BaekjoonHub

+++ 프로그래머스/unrated/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.java
@@ -1,7 +1,7 @@
 class Solution {
     public int solution(String t, String p) {
         int answer = 0;
-            long a = p.length();
+            int a = p.length();
             for (int i = 0; i <= t.length()-a; i++) {
                 String str = "";
                 for (int j = 0; j < p.length(); j++) {
